See description
Fixed this->pool when pool is a static variable.

Added an allocator, but haven't done anything with it yet. It was planned for std::set in an attempt to be able to do something like std::string a = ""; and have that automatically add to the string pool instead, but that failed. This code is supposed to adhere to the standard library, and creating a custom string class to adhere to that isn't necessary but possible in the future.